fatura v0.1.0 Fatura

Este móduco executamos funcoes de fatura.

Functions

criar_fatura(fatura)

Ao receber fatura retorna um array de faturas

## Examples
  iex> Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
  ["Telefone", "Luz", "Agua"]

  iex> Fatura.criar_fatura("bola")
  "bola"  

criar_faturas(faturas, vencimentos)

A função criar_faturas recebe duas listas de faturas e vencimentos e retorna uma estrutura de Fatura.Conta

## Examples

fatura_existe?(faturas, fatura)

Procura sem existe uma fatura em uma lista de faturas

## Examples
  iex> Fatura.fatura_existe?(["Telefone", "Luz", "Agua"],"Luz")
  true

  iex> Fatura.fatura_existe?(["Telefone", "Luz", "Agua"],"internet")
  false

hello()

Retorna a frase de boas vindas para o usuário

## Examples
  iex> Fatura.hello()
  "Bem Vindo Usuario"

loadfile(filename)

Quando encontrar uma fatura
  ## Examples
    iex>  arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
    iex>  Fatura.save("bill_to_pay",arquivo)
    :ok
    iex>  Fatura.loadfile("bill_to_pay")
    ["Telefone", "Luz", "Agua"]

    iex>  arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
    iex>  Fatura.save("bill_to_pay",arquivo)
    :ok
    iex>  Fatura.loadfile("filenotexist")
    "File no found"

ordenar_fatura(faturas)

Ao receber uma lista de faturas, entrega ordenada.

## Examples
  iex> Fatura.ordenar_fatura(["Telefone", "Luz", "Agua"])
  ["Agua", "Luz", "Telefone"]

save(filename, faturas)

Exemplo de criação de arquivo com uma lista dentro

## Examples
  iex> arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
  iex> Fatura.save("bill_to_pay",arquivo)
  :ok