fatura v0.1.0 Fatura

Este móduco executamos funcoes de fatura.

Functions

criar_fatura(var)

Ao receber fatura retorna um array de faturas

## Examples
  iex> Fatura.criar_fatura(["Telefone", "Luz", "Agua "])
  ["Telefone", "Luz", "Agua "]

  iex> Fatura.criar_fatura("bola")
  "bola"  

fatura_existe?(faturas, fatura)

Procura sem existe uma fatura em uma lista de faturas

## Examples

iex> Fatura.fatura_existe?(["Telefone", "Luz", "Agua"],"Luz")
true

iex> Fatura.fatura_existe?(["Telefone", "Luz", "Agua"],"internet")
false

hello()

Retorna a frase de boas vindas para o usuário

## Examples
  iex> Fatura.hello()
  "Bem Vindo Usuario"

ordenar_fatura(faturas)

Ao receber uma lista de faturas, entrega ordenada.

## Examples
  iex> Fatura.ordenar_fatura(["Telefone", "Luz", "Agua"])
  ["Agua", "Luz", "Telefone"]