fatura v0.1.0 Fatura

Este móduco executamos funcoes de fatura.

Functions

acess_bill_by_index(bills, index)

Access list bills by index

##Example
  iex> Fatura.acess_bill_by_index(Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9]),3)
  {%Fatura.Conta{fatura: "Telefone", vencimento: 9}, 3}
  iex> Fatura.acess_bill_by_index(Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9]),2)
  {%Fatura.Conta{fatura: "Agua", vencimento: 5}, 2}

criar_fatura(fatura)

Ao receber fatura retorna um array de faturas

## Examples
  iex> Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
  ["Telefone", "Luz", "Agua"]

  iex> Fatura.criar_fatura("bola")
  "bola"  

criar_faturas(faturas, vencimentos)

A função criar_faturas recebe duas listas de faturas e vencimentos e retorna uma estrutura de Fatura.Conta

## Examples
  Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9])
  [
  %Fatura.Conta{fatura: "Telefone", vencimento: 5},
  %Fatura.Conta{fatura: "Luz", vencimento: 5},
  %Fatura.Conta{fatura: "Agua", vencimento: 5},
  %Fatura.Conta{fatura: "Telefone", vencimento: 9},
  %Fatura.Conta{fatura: "Luz", vencimento: 9},
  %Fatura.Conta{fatura: "Agua", vencimento: 9}
  ]

fatura_existe?(faturas, fatura)

Procura sem existe uma fatura em uma lista de faturas

## Examples
  iex> Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9]) |> Fatura.fatura_existe?(%Fatura.Conta{fatura: "Agua", vencimento: 9})
  true
  iex> fatura = Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9])
  iex> Fatura.fatura_existe?(fatura, %Fatura.Conta{fatura: "Telefone", vencimento: 5} )
  true

hello()

Retorna a frase de boas vindas para o usuário

## Examples
  iex> Fatura.hello()
  "Bem Vindo Usuario"

loadfile(filename)

Quando encontrar uma fatura

## Examples
    iex>  arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
    iex>  Fatura.save("bill_to_pay",arquivo)
    :ok
    iex>  Fatura.loadfile("bill_to_pay")
    ["Telefone", "Luz", "Agua"]

    iex>  arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
    iex>  Fatura.save("bill_to_pay",arquivo)
    :ok
    iex>  Fatura.loadfile("filenotexist")
    "File no found"

pay_bill(bill, index)

Remove a fatura passada no index Remove bill past in index

## Exemples
  iex>Enum.split([1,2,3,4,5],1)
  {[1], [2, 3, 4, 5]}

printbinaryfile(file)

Receive `file_name` and print all data.
  ## Exemples

save(filename, faturas)

Exemplo de criação de arquivo com uma lista dentro

## Examples
  iex> arquivo=Fatura.criar_fatura(["Telefone", "Luz", "Agua"])
  iex> Fatura.save("bill_to_pay",arquivo)
  :ok

sort_bill(bill)

Ao receber uma lista de faturas, entrega ordenada.

  ## Examples
    iex> listbill = Fatura.criar_faturas(["Telefone", "Luz", "Agua"], [5,9])      
    iex> Fatura.sort_bill(listbill)
    [
      %Fatura.Conta{fatura: "Agua", vencimento: 5},
      %Fatura.Conta{fatura: "Agua", vencimento: 9},
      %Fatura.Conta{fatura: "Luz", vencimento: 5},
      %Fatura.Conta{fatura: "Luz", vencimento: 9},
      %Fatura.Conta{fatura: "Telefone", vencimento: 5},
      %Fatura.Conta{fatura: "Telefone", vencimento: 9}
    ]